+++ src/platform_impl/macos/appkit/window.rs
@@ -168,6 +168,9 @@ extern_methods!(
         #[sel(setLevel:)]
         pub fn setLevel(&self, level: NSWindowLevel);
 
+        #[sel(setDocumentEdited:)]
+        pub fn setDocumentEdited(&self, val: bool);
+
         #[sel(occlusionState)]
         pub fn occlusionState(&self) -> NSWindowOcclusionState;
 
@@ -185,6 +188,9 @@ extern_methods!(
 
         #[sel(isZoomed)]
         pub fn isZoomed(&self) -> bool;
+
+        #[sel(isDocumentEdited)]
+        pub fn isDocumentEdited(&self) -> bool;
 
         #[sel(close)]
         pub fn close(&self);